spin off packages.hpp to avoid circular dependencies
