+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 #=========================================================================================
-# (C) (or copyright) 2020-2021. Triad National Security, LLC. All rights reserved.
+# (C) (or copyright) 2020-2022. Triad National Security, LLC. All rights reserved.
 #
 # This program was produced under U.S. Government contract 89233218CNA000001 for Los
 # Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
@@ -135,6 +135,7 @@ add_library(parthenon
   interface/swarm_device_context.hpp
   interface/metadata.cpp
   interface/metadata.hpp
+  interface/packages.hpp
   interface/params.hpp
   interface/sparse_pack.hpp
   interface/sparse_pack_base.cpp
